backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.6.1
python-dotenv==1.0.0
sendgrid==6.10.0
gunicorn==21.2.0
python-dateutil==2.8.2
APScheduler==3.10.4
dnspython==2.4.2
email-validator==2.1.0
Werkzeug==2.3.7
Pillow==10.0.0
pytest==7.3.1
requests==2.28.2
PyJWT==2.6.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
graphql-server==3.0.0b5
python-magic==0.4.27
redis==5.0.1
Flask-Limiter==3.5.0
marshmallow==3.20.1
boto3==1.34.0
cryptography==41.0.3
aiohttp==3.9.1
sentry-sdk==1.39.1
blinker==1.6.3 